DK64: centralize map and setup parsing
